+++ CHANGELOG.md
@@ -7,6 +7,16 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## Changes
+
+- Using new `internal/dialer` package to handle dialing for Async and Sync connections with automatic retires
+- Handling proper backoffs for accept loop in `Server` so server does not just crash when many connections are opened at
+  once
+
+## Fixes
+
+- Fixing `SetBaseContext`, `SetOnClosed`, and `SetPreWrite` functions to not error out if a valid function is used
+
 ## [v0.4.0] - 2022-03-24 (Beta)
 
 ## Changes