+++ CHANGELOG.md
@@ -11,7 +11,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Using new `internal/dialer` package to handle dialing for Async and Sync connections with automatic retires
 - Handling proper backoffs for accept loop in `Server` so server does not just crash when many connections are opened at
-  once
+  once r
 
 ## Fixes