+++ CHANGELOG.md
@@ -16,6 +16,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## Fixes
 
 - Fixing `SetBaseContext`, `SetOnClosed`, and `SetPreWrite` functions to not error out if a valid function is used
+- Async test cases are less flaky
 
 ## [v0.4.0] - 2022-03-24 (Beta)