+++ CHANGELOG.md
@@ -7,11 +7,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [v0.4.1] - 2022-03-24 (Beta)
+
 ## Changes
 
 - Using new `internal/dialer` package to handle dialing for Async and Sync connections with automatic retires
 - Handling proper backoffs for accept loop in `Server` so server does not just crash when many connections are opened at
-  once r
+  once
 
 ## Fixes
 
@@ -207,16 +209,28 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 Initial Release of Frisbee
 
-[unreleased]: https://github.com/loopholelabs/frisbee/compare/v0.4.0...HEAD
+[unreleased]: https://github.com/loopholelabs/frisbee/compare/v0.4.1...HEAD
+
+[v0.4.1]: https://github.com/loopholelabs/frisbee/compare/v0.4.0...v0.4.1
+
 [v0.4.0]: https://github.com/loopholelabs/frisbee/compare/v0.3.2...v0.4.0
+
 [v0.3.2]: https://github.com/loopholelabs/frisbee/compare/v0.3.1...v0.3.2
+
 [v0.3.1]: https://github.com/loopholelabs/frisbee/compare/v0.3.0...v0.3.1
+
 [v0.3.0]: https://github.com/loopholelabs/frisbee/compare/v0.2.4...v0.3.0
+
 [v0.2.4]: https://github.com/loopholelabs/frisbee/compare/v0.2.3...v0.2.4
+
 [v0.2.3]: https://github.com/loopholelabs/frisbee/compare/v0.2.2...v0.2.3
+
 [v0.2.2]: https://github.com/loopholelabs/frisbee/compare/v0.2.1...v0.2.2
+
 [v0.2.1]: https://github.com/loopholelabs/frisbee/compare/v0.2.0...v0.2.1
+
 [v0.2.0]: https://github.com/loopholelabs/frisbee/compare/v0.1.6...v0.2.0
+
 [v0.1.6]: https://github.com/loopholelabs/frisbee/compare/v0.1.5...v0.1.6
 [v0.1.5]: https://github.com/loopholelabs/frisbee/compare/v0.1.4...v0.1.5
 [v0.1.4]: https://github.com/loopholelabs/frisbee/compare/v0.1.3...v0.1.4